+++ gpx2csv/src/main.rs
@@ -140,10 +140,6 @@ fn main() {
     std::process::exit(match run() {
         Ok(_) => 0, // everying is hunky dory
         Err(err) => {
-            Builder::new()
-                .filter_level(LevelFilter::Error)
-                .target(Target::Stdout)
-                .init();
             log::error!("{}", err.to_string().replace("\"", ""));
             1 // exit with a non-zero return code, indicating a problem
         }