+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+### Features
+
+- Start on fitrename
+
 ### Miscellaneous Tasks
 
 - Changelog update